v2.0

+++ RegistroSemijoias/Models/Semijoias.cs
@@ -5,7 +5,7 @@
         public int Id { get; set; }
         public string Nome { get; set; }
         public string Descricao { get; set; }
-        public decimal Preco { get; set; }
+        public double Preco { get; set; }
         public string ImagemUrl { get; set; }
         public int Estoque { get; set; }
         public string Categoria { get; set; }
